Factor out FileSource file add/remove actions
