+++ tests/Functional/Controller/UserSourceControllerTest.php
@@ -18,13 +18,11 @@ use App\Request\GitSourceRequest;
 use App\Request\SourceRequestInterface;
 use App\Services\RunSourceSerializer;
 use App\Services\Source\Store;
-use App\Tests\Model\Route;
 use App\Tests\Model\UserId;
 use App\Tests\Services\AuthorizationRequestAsserter;
 use App\Tests\Services\EntityRemover;
 use App\Tests\Services\FileStoreFixtureCreator;
 use App\Tests\Services\FixtureLoader;
-use App\Validator\YamlFilenameConstraint;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
 
@@ -494,194 +492,5 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
         self::assertSame($expectedResponse->getStatusCode(), $response->getStatusCode());
         self::assertSame($expectedResponse->headers->get('content-type'), $response->headers->get('content-type'));
         self::assertSame($expectedResponse->getContent(), $response->getContent());
-    }
-
-    /**
-     * @dataProvider addFileInvalidRequestDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testAddFileInvalidRequest(
-        SourceInterface $source,
-        string $userId,
-        string $filename,
-        string $content,
-        array $expectedResponseData
-    ): void {
-        $this->store->add($source);
-
-        $this->setUserServiceAuthorizedResponse($userId);
-
-        $response = $this->applicationClient->makeAuthorizedRequest(
-            'POST',
-            new Route('add_file', [
-                'sourceId' => $source->getId(),
-                'filename' => $filename,
-            ]),
-            [
-                'content' => $content,
-            ]
-        );
-
-        self::assertSame(400, $response->getStatusCode());
-        $this->authorizationRequestAsserter->assertAuthorizationRequestIsMade();
-        self::assertInstanceOf(JsonResponse::class, $response);
-
-        $responseData = json_decode((string) $response->getContent(), true);
-        self::assertEquals($expectedResponseData, $responseData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function addFileInvalidRequestDataProvider(): array
-    {
-        $userId = UserId::create();
-        $label = 'file source label';
-
-        $fileSource = new FileSource($userId, $label);
-
-        $expectedInvalidFilenameResponseData = $this->createExpectedInvalidFilenameResponseData();
-
-        return [
-            'name empty with .yaml extension, content non-empty' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => '.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-            'name contains backslash characters, content non-empty' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'one two \\ three.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-            'name contains null byte characters, content non-empty' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'one ' . chr(0) . ' two three' . chr(0) . '.yaml',
-                'content' => 'non-empty value',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-            'name valid, content empty' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'filename.yaml',
-                'content' => '',
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'content' => [
-                                'value' => '',
-                                'message' => 'File content must not be empty.',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            'name valid, content invalid yaml' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'filename.yaml',
-                'content' => "- item\ncontent",
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'content' => [
-                                'value' => '',
-                                'message' => 'Content must be valid YAML: Unable to parse at line 2 (near "content").',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider removeFileInvalidRequestDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testRemoveFileInvalidRequest(
-        SourceInterface $source,
-        string $userId,
-        string $filename,
-        array $expectedResponseData
-    ): void {
-        $this->store->add($source);
-
-        $this->setUserServiceAuthorizedResponse($userId);
-
-        $response = $this->applicationClient->makeAuthorizedRequest(
-            'DELETE',
-            new Route('remove_file', [
-                'sourceId' => $source->getId(),
-                'filename' => $filename,
-            ])
-        );
-
-        self::assertSame(400, $response->getStatusCode());
-        $this->authorizationRequestAsserter->assertAuthorizationRequestIsMade();
-        self::assertInstanceOf(JsonResponse::class, $response);
-
-        $responseData = json_decode((string) $response->getContent(), true);
-        self::assertEquals($expectedResponseData, $responseData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function removeFileInvalidRequestDataProvider(): array
-    {
-        $userId = UserId::create();
-        $label = 'file source label';
-
-        $fileSource = new FileSource($userId, $label);
-
-        $expectedInvalidFilenameResponseData = $this->createExpectedInvalidFilenameResponseData();
-
-        return [
-            'name empty with .yaml extension' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => '.yaml',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-            'name contains backslash characters' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'one two \\ three.yaml',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-            'name contains null byte characters' => [
-                'source' => $fileSource,
-                'userId' => $userId,
-                'filename' => 'one ' . chr(0) . ' two three' . chr(0) . '.yaml',
-                'expectedResponseData' => $expectedInvalidFilenameResponseData,
-            ],
-        ];
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    private function createExpectedInvalidFilenameResponseData(): array
-    {
-        return [
-            'error' => [
-                'type' => 'invalid_request',
-                'payload' => [
-                    'name' => [
-                        'value' => '',
-                        'message' => YamlFilenameConstraint::MESSAGE_NAME_INVALID,
-                    ],
-                ],
-            ],
-        ];
     }
 }
